Added Version number to About table view
Added a row to the About screen table view to display the current
version of the App.  I thought it would be useful for users and
developers to know which version of the App they are using.

Please review the method I used to obtain the version number and if it
is acceptable pull this into the master branch.

Shadallark