iPad Storyboard created
I copied everything from the iPhone storyboard and pasted it into the
iPad storyboard (reorganized a bit to make it easier to follow).

Adjusted the project settings to indicate that the app could be built
for iPhone or iPad (ie. Universal).

iPad version allows the app to be rotated in any direction.