Add notification observer to check orientation before setting screen size for HRPGBaseViewController
As recommended by @vIiRuS I added a notification observer to the
HRPGBaseViewController for UIDeviceOrientationDidChangeNotification.

In the orientationChanged function the code checks the orientation of
the User Interface and sets the screenWidth accordingly (=
screenRect.size.width for Portrait; = screenRect.size.height for
Landscape).  This way the proper width is returned for the iPhone
(Portrait only), for the iPad (any orientation), or for when rotation
is locked.