Fixed the instantiation of the FontAwesomeIconFactory
Figured out where my issue was in creating the instance of the
FontAwesomeIconFactory.

Now the icons and streak count show up correctly.  They are both grey
if the streak count is 0 or green if the streak count is greater than 0
(let me know if you would prefer a different colour scheme).