Height and Width constraints on streakImage view and content hugging on streakLabel
As suggested by @vIiRuS I have added constraints to the streakImage
view and adjusted the properties of the streakLabel to allow it to
resize with the content.

Instead of fixing the height directly of the streakImage I told it to
maintain an aspect ratio of 1:1 to the width as that allows the icons
to display as expected.